move default velocity secret warning to PhantazmServer startup
